Reorganize feign package
