--- src/main/java/com/iexec/worker/feign/CoreClient.java
@@ -1,14 +1,0 @@
-package com.iexec.worker.feign;
-
-import feign.FeignException;
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@FeignClient(name = "CoreClient",
-        url = "${core.protocol}://${core.host}:${core.port}",
-        configuration = FeignConfiguration.class)
-public interface CoreClient {
-
-    @GetMapping("/version")
-    String getCoreVersion() throws FeignException;
-}
