+++ src/main/java/com/iexec/worker/feign/client/ResultClient.java
@@ -1,7 +1,9 @@
-package com.iexec.worker.feign;
+package com.iexec.worker.feign.client;
 
 import com.iexec.common.result.ResultModel;
 import com.iexec.common.result.eip712.Eip712Challenge;
+import com.iexec.worker.feign.config.FeignConfiguration;
+
 import feign.FeignException;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.ResponseEntity;
@@ -11,10 +13,10 @@ import org.springframework.web.bind.annotation.*;
 @FeignClient(name = "ResultRepoClient",
         url = "#{publicConfigurationService.resultRepositoryURL}",
         configuration = FeignConfiguration.class)
-public interface ResultRepoClient {
+public interface ResultClient {
 
     @GetMapping("/results/challenge")
-    Eip712Challenge getChallenge(@RequestParam(name = "chainId") Integer chainId) throws FeignException;
+    ResponseEntity<Eip712Challenge> getChallenge(@RequestParam(name = "chainId") Integer chainId) throws FeignException;
 
     @PostMapping("/results")
     ResponseEntity<String> uploadResult(@RequestHeader("Authorization") String customToken,
